+++ src/components/Movies/Movie.jsx
@@ -1,12 +1,10 @@
 import { useEffect } from "react";
 import { getAllData } from "../helpers/get";
 import { useState } from "react";
-// import movieLogo from "../../assets/icon-category-movie.svg";
-import seriesLogo from "../../assets/icon-category-tv.svg";
-
-function Serie() {
+import movieLogo from "../../assets/icon-category-movie.svg";
+function Movie() {
   const [movies, setMovies] = useState([]);
-  console.log(movies);
+
   const getMovies = async () => {
     const movies = await getAllData();
     setMovies(movies);
@@ -16,20 +14,17 @@ function Serie() {
   }, []);
   return (
     <>
-    
       {movies.map((movie, index) => {
-        if (movie.category==="TV Series") {
+        if (movie.category === "Movie") {
           return (
             <div key={index}>
-             
               <img
                 src={"src" + movie.thumbnail.regular.small.substring(1)}
                 alt={movie.title + "'s image"}
               />
               <p>{movie.year}</p>
-              {console.log(movie.category)}
               <p style={{ backgroundColor: "#999999" }}>
-                <img src={seriesLogo} />
+                <img src={movieLogo} />
                 {movie.category}
               </p>
               <p>{movie.rating}</p>
@@ -42,4 +37,4 @@ function Serie() {
     </>
   );
 }
-export default Serie;
+export default Movie;
