+++ src/components/TvSeries/SerieSet.jsx
@@ -1,13 +1,11 @@
-
-
 import { useEffect } from "react";
-import { getAllData } from "../components/helpers/get";
+import { getAllData } from "../helpers/get";
 import { useState } from "react";
-import movieLogo from "../assets/icon-category-movie.svg";
-import seriesLogo from "../assets/icon-category-tv.svg";
-function Movie() {
+import seriesLogo from "../../assets/icon-category-tv.svg";
+
+function SerieSet() {
   const [movies, setMovies] = useState([]);
-  console.log(movies);
+
   const getMovies = async () => {
     const movies = await getAllData();
     setMovies(movies);
@@ -17,20 +15,18 @@ function Movie() {
   }, []);
   return (
     <>
-    
       {movies.map((movie, index) => {
-        if (movie.category==="Movie") {
+        if (movie.category === "TV Series") {
           return (
             <div key={index}>
-             
               <img
                 src={"src" + movie.thumbnail.regular.small.substring(1)}
                 alt={movie.title + "'s image"}
               />
               <p>{movie.year}</p>
-              {console.log(movie.category)}
+
               <p style={{ backgroundColor: "#999999" }}>
-                <img src={movieLogo} />
+                <img src={seriesLogo} />
                 {movie.category}
               </p>
               <p>{movie.rating}</p>
@@ -43,4 +39,4 @@ function Movie() {
     </>
   );
 }
-export default Movie;
+export default SerieSet;
